embedded Test

+++ src/main/java/com/exalt/partssystem/service/CompanyServiceImpl.java
@@ -1,6 +1,7 @@
 package com.exalt.partssystem.service;
 
-import com.exalt.partssystem.error.NotFoundExceptions;
+import com.exalt.partssystem.error.exceptions.ConflictExceptions;
+import com.exalt.partssystem.error.exceptions.NotFoundExceptions;
 import com.exalt.partssystem.model.Company;
 import com.exalt.partssystem.repository.CompanyRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,7 +15,9 @@ import java.util.List;
 
 @Service
 public class CompanyServiceImpl implements CompanyService {
-
+    /*
+    To inject the object companyRepository, implicitly
+     */
     @Autowired
     private CompanyRepository companyRepository;
 
@@ -29,15 +32,22 @@ public class CompanyServiceImpl implements CompanyService {
     @Override
     @Transactional
     public Company get(String name) {
-        return companyRepository.findByName(name);
+        Company company = companyRepository.findByName(name);
+        if (company != null) {
+            return company;
+        } else {
+            throw new NotFoundExceptions("Company not found");
+        }
     }
 
     @Override
-    public Company save(Company company) {
-        companyRepository.deleteByName("test1");
-        if (company.getName().equals("test2"))
-            throw new NotFoundExceptions("sdfsdf");
-        return companyRepository.save(company);
+    public Company save(Company newCompany) {
+        Company company = companyRepository.deleteByName(newCompany.getName());
+        if (company == null) {
+            return companyRepository.save(newCompany);
+        } else {
+            throw new ConflictExceptions(String.format("There is company with the same name ( %s ) ", company.getName()));
+        }
     }
 
     @Override
@@ -48,6 +58,11 @@ public class CompanyServiceImpl implements CompanyService {
 
     @Override
     public Company delete(String name) {
-        return companyRepository.deleteByName(name);
+        Company company = companyRepository.findByName(name);
+        if (company != null) {
+            return companyRepository.deleteByName(name);
+        } else {
+            throw new NotFoundExceptions("Not found company to deleting ");
+        }
     }
 }
